% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateRandomTrendDataIndex.R
\name{generateRandomTrendIndexData}
\alias{generateRandomTrendIndexData}
\alias{getTrendIndexData}
\title{Generate random trend index data}
\usage{
generateRandomTrendIndexData()

getTrendIndexData(searchEngine)
}
\arguments{
\item{searchEngine}{The name of a search engine as a character string.}
}
\description{
A data frame with the following columns.
\itemize{
\item \code{type} (\code{character})
\item \code{date} (\code{Date})
\item \code{close} (\code{numeric})
}
}
