% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appFunctions.R
\name{getUi}
\alias{getUi}
\alias{getServer}
\title{Get ui and server functions}
\usage{
getUi(searchEngines, theme = "lumen")

getServer(searchEngines)
}
\arguments{
\item{searchEngines}{Names of the search engines of interest as a character
vector.}

\item{theme}{Name of a Shiny theme as listed by \link[shinythemes:shinythemes]{shinythemes::shinythemes}.}
}
\description{
Build a ui & server pair for \code{\link[shiny:shinyApp]{shiny::shinyApp()}}.
}
