% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexTrendModule.R
\name{indexTrendModuleUI}
\alias{indexTrendModuleUI}
\alias{indexTrendModuleServer}
\title{Show the trend index time series as a Shiny module}
\usage{
indexTrendModuleUI(id, searchEngine, choices)

indexTrendModuleServer(id, trend_data)
}
\arguments{
\item{id}{The identifier string linking the module functions.}

\item{searchEngine}{Name of the search engines of interest as a character
string.}

\item{choices}{The name of the trend indices of interest.}

\item{trend_data}{A data frame with the following columns.
\itemize{
\item \code{type} (\code{character})
\item \code{date} (\code{Date})
\item \code{close} (\code{numeric})
}}
}
\description{
\code{\link[=indexTrendModuleUI]{indexTrendModuleUI()}} returns a \code{\link[shiny:reexports]{shiny::tagList()}} and \code{\link[=indexTrendModuleServer]{indexTrendModuleServer()}}
returns nothing.
}
