% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pauseAndThinkAboutIt.R
\name{pauseAndThinkAboutIt}
\alias{pauseAndThinkAboutIt}
\title{Suspend Execution for a Time Interval}
\usage{
pauseAndThinkAboutIt(seconds)
}
\arguments{
\item{seconds}{A number of seconds}
}
\value{
Invisible \code{NULL}.
}
\description{
These are descriptively renamed wrappers of \code{\link[base:Sys.sleep]{base::Sys.sleep()}}
}
