% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTrendIndex.R
\name{plotTrendIndex}
\alias{plotTrendIndex}
\title{Plot trend index}
\usage{
plotTrendIndex(date, close, smoother, smootherSpan)
}
\arguments{
\item{date}{A \code{Date} vector of same length as \code{close}.}

\item{close}{Index value at close time as a \code{numeric} vector of same length as \code{date}.}

\item{smoother}{Should a \code{\link[stats:lowess]{stats::lowess()}} smoothed version of the trend index
be plotted as well? Either \code{TRUE} and \code{FALSE}.}

\item{smootherSpan}{the smoother span. This gives the proportion of points in
the plot which influence the smooth at each value. Larger values give more
smoothness.}
}
\value{
Nothing silently.
}
\description{
Plot trend index as a time series and optionally a smoothed version.
}
