% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runInternalApp.R
\name{runInternalApp}
\alias{runInternalApp}
\alias{getInternalAppNames}
\alias{getInternalAppFolder}
\title{Start an internal shiny app}
\usage{
runInternalApp(appName, ...)

getInternalAppNames()

getInternalAppFolder(appName = "")
}
\arguments{
\item{appName}{The name as a single character string.}

\item{...}{Passed to \code{\link[shiny:runApp]{shiny::runApp()}}}
}
\value{
\code{\link[=getInternalAppFolder]{getInternalAppFolder()}} returns a folder path as a character string.
\code{\link[=getInternalAppNames]{getInternalAppNames()}} returns app names as a character vector.
\code{\link[=runInternalApp]{runInternalApp()}} run the apps and returns nothing.
}
\description{
Locate internal shiny apps, list their names and run one.
}
