% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showPresentation.R
\name{showPresentation}
\alias{showPresentation}
\alias{getPresentationFolder}
\alias{getPresentationNames}
\title{Show presentation}
\usage{
showPresentation(presentationName = "shinydo-presentation.html")

getPresentationFolder(presentationName = "")

getPresentationNames()
}
\arguments{
\item{presentationName}{The name of the presentation as a character string.}
}
\description{
Locate presentations, list their names and run one in the default web browser.
}
